float

+++ C/aula9/manipulaconver.c
@@ -6,10 +6,10 @@ int main() {
     int soma = a + b;
     int diferenca = a - b;
     int produto = a * b;
-    int quociente = a / b;
+    float quociente = (float)a / b;
 
     printf("Soma: %d\n", soma);
     printf("Diferença: %d\n", diferenca);
     printf("Produto: %d\n", produto);
-    printf("Quociente: %d\n", quociente);
+    printf("Quociente: %.2f\n", quociente);
 }
